+++ TradeRecord.cs
@@ -11,13 +11,13 @@ namespace stockmarket
         internal int month { get; private set; }
         internal int day { get; private set; }
         internal int year { get; private set; }
-        internal double price { get; private set; }
+        internal decimal price { get; private set; }
         internal int shares { get; private set; }
-        internal double money { get; private set; }
+        internal decimal money { get; private set; }
 
         internal TradeRecord(TradeAction type,
             int month, int day, int year,
-            double price, int shares, double money)
+            decimal price, int shares, decimal money)
         {
             this.type = type;
             this.month = month;
